# Conflicts:
#	example-app/app/Http/Controllers/Admin/ControllerProductManager.php
#	example-app/app/Http/Controllers/ControllerGridPage.php
#	example-app/app/Http/Controllers/ControllerView.php
#	example-app/database/migrations/2023_11_15_055508_create_bills_table.php
#	example-app/resources/views/backend/product/table.blade.php
#	example-app/resources/views/fontend/black.blade.php
#	example-app/resources/views/fontend/grid.blade.php
#	example-app/resources/views/fontend/index.blade.php
#	example-app/routes/web.php

+++ example-app/app/Http/Controllers/Admin/ControllerProductManager.php
@@ -25,49 +25,70 @@ class ControllerProductManager extends Controller
     }
     public function create_handler(Request $request)
     {
-        $request->validate([
-            'name' => 'required|string|max:250',
-            'image' =>  'required|image|mimes:png,jpg,jpeg|max:2048',
-            'description' => 'required',
-            'price' => 'required|numeric|gt:0',
-            'inventory' => 'required|numeric|gt:0',
-        ]);
 
-        $image = $request->file('image');
-        $imageName = time() . '.' . $image->getClientOriginalExtension();
+        $token = $request->input('_token');
 
-        $product = new Product([
-            'name' => $request['name'],
-            'image' => $imageName,
-            'categories_id' => $request['cate'],
-            'manufacturer_id' => $request['manu'],
-            'description' => $request['description'],
-            'price' => $request['price'],
-            'inventory' => $request['inventory'],
-        ]);
-        if ($product->save()) {
-            $image->move(public_path('images/products'), $imageName);
-            return redirect()->route('product.table')->with('success', 'Thêm sản phẩm thành công');
+        if (Session::has('_token') && Session::get('_token') === $token) {
+            $request->validate([
+                'name' => 'required|string|max:255',
+                'image' =>  'required|image|mimes:png,jpg,jpeg|max:2048 ',
+                'description' => 'required',
+                'price' => 'required|numeric|gt:0',
+                'inventory' => 'required|numeric|gt:0',
+                'manu' => 'required',
+                'cate' => 'required',
+            ]);
+            if (!!!Categories::find($request['cate'])) {
+                return back()->withErrors(['cate' => 'Category does not exist.']);
+            }
+            if (!!!Manufacturers::find($request['manu'])) {
+                return back()->withErrors(['manu' => 'Manufacturer does not exist.']);
+            }
+            $image = $request->file('image');
+            $imageName = time() . '.' . $image->getClientOriginalExtension();
+
+            $product = new Product([
+                'name' => $request['name'],
+                'image' => $imageName,
+                'categories_id' => $request['cate'],
+                'manufacturer_id' => $request['manu'],
+                'description' => $request['description'],
+                'price' => $request['price'],
+                'inventory' => $request['inventory'],
+                'unique_token' => Str::uuid()->toString(),
+            ]);
+            if ($product->save()) {
+                $image->move(public_path('images/products'), $imageName);
+                return redirect()->route('product.table')->with('success', 'Thêm sản phẩm thành công');
+            }
+            Session::put('token', $token);
+
+            return back();
+        } else {
         }
-        return back();
     }
     public function edit($id)
     {
-        if (!$product = Product::find($id)) {
-            return redirect()->back()->with('errors', 'Danh mục không tồn tại');
-        }
         $product = Product::find($id);
+
+        if (!$product) {
+            return redirect()->route('product.table')->with('errors', 'Danh mục không tồn tại');
+        }
+
         $manus = Manufacturers::all();
         $cates = Categories::all();
         $page = 'Product edit';
+
         return view('backend.product.edit', compact('product', 'page', 'cates', 'manus'));
     }
     public function edit_handle($id, Request $request)
     {
-        if (!$product = Product::find($id)) {
-            return redirect('admin/product/table')->with('errors', 'Danh mục không tồn tại');
-        }
         $product = Product::find($id);
+
+        if (!$product) {
+            return redirect()->route('product.table')->with('errors', 'Danh mục không tồn tại');
+        }
+
         $request->validate([
             'name' => 'required|string|max:255',
             'image' =>  'nullable|image|mimes:png,jpg,jpeg|max:2048',
@@ -102,20 +123,26 @@ class ControllerProductManager extends Controller
     //xóa  sản phẩm:
     public function delete($id)
     {
-        if (!$product = Product::find($id)) {
-            return redirect()->back()->with('errors', 'Product does not exist!');
-        }
-        $oders = Orders::where('product_id', $id)->get();
-        if ($oders->count() == 0) {
-            $product = Product::find($id);
-            $path = "images/" . $product->image;
-            if (File::exists($path)) {
-                File::delete($path);
+        $token = $request->input('_token');
+        if (Session::has('_token') && Session::get('_token') === $token) {
+            $id = $request['id'];
+            if (!$product = Product::find($id)) {
+                return redirect()->back()->with('errors', 'Sản phẩm không tồn tại!');
             }
-            $product->delete();
-            return redirect()->back()->with('success', 'Delete product succeed!');
+            $oders = Orders::where('product_id', $id)->get();
+            if ($oders->count() == 0) {
+                $product = Product::find($id);
+                $path = "images/products/" . $product->image;
+                if (File::exists($path)) {
+                    File::delete($path);
+                }
+                $product->delete();
+                return redirect()->route('product.table')->with('success', 'Xóa thành công');
+            } else {
+                return redirect()->back()->with('errors', 'Xóa thất bại!');
+            }
         } else {
-            return redirect()->back()->with('errors', 'Cannot delete product!');
+            return redirect()->route('product.table')->with('error', 'Không tìm thấy sản phẩm!');
         }
     }
 }
